% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultlist.R
\name{$.defaultlist}
\alias{$.defaultlist}
\title{Fetch value from defaultlist: [[]] or $}
\usage{
\method{$}{defaultlist}(x, y)
}
\arguments{
\item{x}{defaultlist object}

\item{y}{character or numeric index}
}
\description{
Fetch value from defaultlist: [[]] or $
}
