% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultlist.R
\name{convert_list_to_dlist}
\alias{convert_list_to_dlist}
\title{Internal helper to consistently convert a list to a defaultlist}
\usage{
convert_list_to_dlist(lst, .default)
}
\arguments{
\item{lst}{list object}

\item{.default}{default value to return when item not in list}
}
\description{
Internal helper to consistently convert a list to a defaultlist
}
