% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultlist.R
\name{dlist}
\alias{dlist}
\alias{empty_dlist}
\title{Create a \code{defaultlist} object}
\usage{
dlist(..., .default)

empty_dlist(.default)
}
\arguments{
\item{...}{objects with which to initialise the list.}

\item{.default}{default value to return when item not in list}
}
\value{
new `defaultlist` object
}
\description{
Similar to a `defaultdict` in Python, a \code{defaultlist} is mostly identical
to a list except it will return the user-defined default value if the
requested elements are not in the list.
}
