% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultlist.R
\name{print.defaultlist}
\alias{print.defaultlist}
\title{Print defaultlist}
\usage{
\method{print}{defaultlist}(x, ...)
}
\arguments{
\item{x}{defaultlist object}

\item{...}{other arguments ignored}
}
\description{
Drop attributes to ensure it looks like a standard list.
}
